# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorShield API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type SignupRequestUserType =
  (typeof SignupRequestUserType)[keyof typeof SignupRequestUserType];

export const SignupRequestUserType = {
  senior: "senior",
  adult_child: "adult_child",
  staff: "staff",
} as const;

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  user_type: SignupRequestUserType;
  first_name?: string;
  last_name?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user_id: string;
  token: string;
  user_type: string;
  first_name?: string;
  last_name?: string;
  onboarding_completed: boolean;
}

export interface VerifyResponse {
  user_id: string;
  email: string;
  user_type: string;
}

export interface UserProfile {
  id: string;
  email: string;
  first_name?: string;
  last_name?: string;
  phone_number?: string;
  user_type: string;
  onboarding_completed: boolean;
  onboarding_step?: number;
  created_at?: string;
}

export interface UpdateProfileRequest {
  first_name?: string;
  last_name?: string;
  phone_number?: string;
  onboarding_completed?: boolean;
  onboarding_step?: number;
}

export interface UserPreferences {
  preferred_voice: string;
  voice_speed: number;
  voice_volume: number;
  color_scheme: string;
  high_contrast_enabled: boolean;
  font_size: string;
  haptic_feedback: boolean;
  captions_enabled: boolean;
  data_collection_enabled: boolean;
}

export interface UpdatePreferencesRequest {
  preferred_voice?: string;
  voice_speed?: number;
  voice_volume?: number;
  color_scheme?: string;
  high_contrast_enabled?: boolean;
  font_size?: string;
  haptic_feedback?: boolean;
  captions_enabled?: boolean;
  data_collection_enabled?: boolean;
}

export interface VoiceTextRequest {
  request_text: string;
}

export interface VoiceResponse {
  request_text: string;
  response_text: string;
  task_category: string;
  success: boolean;
  error_message?: string;
  history_id?: string;
}

export interface VoiceHistoryItem {
  id: string;
  request_text: string;
  response_text: string;
  task_category: string;
  success: boolean;
  created_at: string;
}

export interface ScamAnalyzeRequest {
  text: string;
}

export type ScamAnalysisResultRiskLevel =
  (typeof ScamAnalysisResultRiskLevel)[keyof typeof ScamAnalysisResultRiskLevel];

export const ScamAnalysisResultRiskLevel = {
  safe: "safe",
  suspicious: "suspicious",
  high_risk: "high_risk",
} as const;

export type ScamAnalysisResultAnalysisDetails = { [key: string]: unknown };

export interface ScamAnalysisResult {
  id: string;
  risk_score: number;
  risk_level: ScamAnalysisResultRiskLevel;
  detected_patterns: string[];
  explanation: string;
  analysis_details?: ScamAnalysisResultAnalysisDetails;
}

export interface ScamHistoryItem {
  id: string;
  extracted_text?: string;
  risk_score: number;
  risk_level: string;
  user_feedback?: string;
  created_at: string;
}

export type ScamFeedbackRequestFeedbackType =
  (typeof ScamFeedbackRequestFeedbackType)[keyof typeof ScamFeedbackRequestFeedbackType];

export const ScamFeedbackRequestFeedbackType = {
  correct: "correct",
  false_positive: "false_positive",
  false_negative: "false_negative",
} as const;

export interface ScamFeedbackRequest {
  scam_analysis_id: string;
  feedback_type: ScamFeedbackRequestFeedbackType;
  explanation?: string;
}

export interface FamilyMember {
  id: string;
  email: string;
  first_name?: string;
  last_name?: string;
  relationship: string;
  scam_alerts: boolean;
  weekly_summary: boolean;
  created_at: string;
}

export interface AddFamilyMemberRequest {
  adult_child_email: string;
  relationship: string;
}

export interface Alert {
  id: string;
  alert_type: string;
  message: string;
  read: boolean;
  created_at: string;
}

export type SubscriptionInfoTier =
  (typeof SubscriptionInfoTier)[keyof typeof SubscriptionInfoTier];

export const SubscriptionInfoTier = {
  free: "free",
  premium_trial: "premium_trial",
  premium_paid: "premium_paid",
} as const;

export interface SubscriptionInfo {
  tier: SubscriptionInfoTier;
  status: string;
  billing_cycle?: string;
  trial_end_date?: string;
  premium_end_date?: string;
}

export type CheckoutRequestBillingCycle =
  (typeof CheckoutRequestBillingCycle)[keyof typeof CheckoutRequestBillingCycle];

export const CheckoutRequestBillingCycle = {
  monthly: "monthly",
  annual: "annual",
} as const;

export interface CheckoutRequest {
  tier: string;
  billing_cycle: CheckoutRequestBillingCycle;
}

export interface CheckoutResponse {
  checkout_url: string;
}

export interface SupportTicket {
  id: string;
  subject: string;
  message: string;
  status: string;
  priority: string;
  response?: string;
  created_at: string;
  updated_at?: string;
}

export interface CreateTicketRequest {
  subject: string;
  message: string;
}

export interface SupportTicketResponse {
  success: boolean;
  ticket_id: string;
  status: string;
}

export interface FaqItem {
  id: string;
  question: string;
  answer: string;
  category: string;
}

export interface AdminMetrics {
  total_users: number;
  free_users: number;
  trial_users: number;
  paid_users: number;
  daily_active_users?: number;
  voice_requests_count?: number;
  scams_detected?: number;
  family_alerts_sent?: number;
}

export type GetVoiceHistory200 = {
  history: VoiceHistoryItem[];
};

export type GetScamHistory200 = {
  history: ScamHistoryItem[];
};

export type GetFamilyMembers200 = {
  members: FamilyMember[];
};

export type GetAlerts200 = {
  alerts: Alert[];
};

export type UpdateAlertBody = {
  read: boolean;
};

export type GetSupportTickets200 = {
  tickets: SupportTicket[];
};

export type GetFaq200 = {
  items: FaqItem[];
};