# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SeniorShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    get:
      operationId: verifyToken
      tags: [auth]
      summary: Verify JWT token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /family/members:
    get:
      operationId: getFamilyMembers
      tags: [family]
      summary: Get linked family members
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Family members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/FamilyMember"
                required:
                  - members
  /family/add-member:
    post:
      operationId: addFamilyMember
      tags: [family]
      summary: Add a family member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFamilyMemberRequest"
      responses:
        "200":
          description: Family member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /family/member/{memberId}:
    delete:
      operationId: removeFamilyMember
      tags: [family]
      summary: Remove a family member
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get all alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                required:
                  - alerts
  /alerts/{alertId}:
    put:
      operationId: updateAlert
      tags: [alerts]
      summary: Mark alert as read
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                read:
                  type: boolean
              required:
                - read
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get user subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionInfo"
  /billing/create-checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create Stripe checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        user_type:
          type: string
          enum: [senior, adult_child, staff]
        first_name:
          type: string
        last_name:
          type: string
      required:
        - email
        - password
        - user_type
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user_id:
          type: string
        token:
          type: string
        user_type:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        onboarding_completed:
          type: boolean
      required:
        - user_id
        - token
        - user_type
        - onboarding_completed
    VerifyResponse:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        user_type:
          type: string
      required:
        - user_id
        - email
        - user_type
...
    FamilyMember:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        relationship:
          type: string
        scam_alerts:
          type: boolean
        weekly_summary:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - email
        - relationship
        - scam_alerts
        - weekly_summary
        - created_at
    AddFamilyMemberRequest:
      type: object
      properties:
        adult_child_email:
          type: string
          format: email
        relationship:
          type: string
      required:
        - adult_child_email
        - relationship
    Alert:
      type: object
      properties:
        id:
          type: string
        alert_type:
          type: string
        message:
          type: string
        read:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - alert_type
        - message
        - read
        - created_at
    SubscriptionInfo:
      type: object
      properties:
        tier:
          type: string
          enum: [free, premium_trial, premium_paid]
        status:
          type: string
        billing_cycle:
          type: string
        trial_end_date:
          type: string
        premium_end_date:
          type: string
      required:
        - tier
        - status
    CheckoutRequest:
      type: object
      properties:
        tier:
          type: string
        billing_cycle:
          type: string
          enum: [monthly, annual]
      required:
        - tier
        - billing_cycle
    CheckoutResponse:
      type: object
      properties:
        checkout_url:
          type: string
      required:
        - checkout_url
...