# lib/api-zod/src/generated/types/verifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SeniorShield API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyResponse {
  user_id: string;
  email: string;
  user_type: string;
}